Reorganize state machine module (#28)
* start by moving state_machine mod into a subdir

* move coordinator and signer into subdirs

* fmt fixes since we've reduced the nesting

* move existing Coordinator info a frost module since we will also have fire/faire/roast
